requirements.txt
langchain
transformers
faiss-cpu
pinecone-client
fastapi
uvicorn
pydantic
python-multipart
pdfplumber
python-docx
pymupdf
sentence-transformers